pyproject.toml
[project]
name = "heimdex_mvp"
version = "0.1.0"
description = "VNA dev environment with ffmpeg/ffprobe, PySceneDetect, and uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "pydantic>=2.8.0",
  "scenedetect>=0.6.3",
  "opencv-python-headless>=4.10.0",
  "ffmpeg-python>=0.2.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "rich>=13.7.1",
  "uvicorn[standard]>=0.31.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.19.0",
  "redis>=5.0.0",
  "rq>=1.15.1",
  "structlog>=24.1.0",
  "alembic>=1.13.2",
  "pydantic-settings>=2.4.0",
  "pyjwt>=2.8.0",
  "tenacity>=9.0.0"
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov",
  "httpx",
  "fakeredis>=2.23.0"
]

[tool.uv]
# uv will create and manage a venv automatically when using `uv sync` or `uv run`.
# No special config required, but this block is here if you want to extend later.

[tool.pyscenedetect]
# Example default options you might apply later.
# min-scene-len = 15
